+++ src/redux/studio.js
@@ -3,8 +3,6 @@ const keyMirror = require('keymirror');
 const api = require('../lib/api');
 const log = require('../lib/log');
 
-const {selectUserId, selectIsAdmin, selectIsSocial} = require('./session');
-
 const Status = keyMirror({
     FETCHED: null,
     NOT_FETCHED: null,
@@ -20,7 +18,6 @@ const getInitialState = () => ({
     commentingAllowed: false,
     thumbnail: '',
     followers: 0,
-    owner: null,
 
     rolesStatus: Status.NOT_FETCHED,
     manager: false,
@@ -58,8 +55,6 @@ const studioReducer = (state, action) => {
     }
 };
 
-// Action Creators
-
 const setFetchStatus = (fetchType, fetchStatus, error) => ({
     type: 'SET_FETCH_STATUS',
     fetchType,
@@ -77,8 +72,6 @@ const setRoles = roles => ({
     roles: roles
 });
 
-// Thunks
-
 const getInfo = studioId => (dispatch => {
     dispatch(setFetchStatus('infoStatus', Status.FETCHING));
     api({uri: `/studios/${studioId}`}, (err, body, res) => {
@@ -93,8 +86,7 @@ const getInfo = studioId => (dispatch => {
             openToAll: body.open_to_all,
             commentingAllowed: body.commenting_allowed,
             updated: new Date(body.history.modified),
-            followers: body.stats.followers,
-            owner: body.owner
+            followers: body.stats.followers
         }));
     });
 });
@@ -119,34 +111,10 @@ const getRoles = (studioId, username, token) => (dispatch => {
     });
 });
 
-// Selectors
-
-// Fine-grain selector helpers - not exported, use the higher level selectors below
-const isCreator = state => selectUserId(state) === state.studio.owner;
-const isCurator = state => state.studio.curator;
-const isManager = state => state.studio.manager || isCreator(state);
-
-// Action-based permissions selectors
-const selectCanEditInfo = state => selectIsAdmin(state) || isManager(state);
-const selectCanAddProjects = state =>
-    isManager(state) ||
-    isCurator(state) ||
-    (selectIsSocial(state) && state.studio.openToAll);
-
-// This isn't "canComment" since they could be muted, but comment composer handles that
-const selectShowCommentComposer = state => selectIsSocial(state);
-
 module.exports = {
     getInitialState,
     studioReducer,
     Status,
-
-    // Thunks
     getInfo,
-    getRoles,
-
-    // Selectors
-    selectCanEditInfo,
-    selectCanAddProjects,
-    selectShowCommentComposer
+    getRoles
 };
