+++ src/views/components/components.jsx
@@ -8,78 +8,22 @@ const Carousel = require('../../components/carousel/carousel.jsx');
 const Form = require('../../components/forms/form.jsx');
 const Input = require('../../components/forms/input.jsx');
 const Spinner = require('../../components/spinner/spinner.jsx');
-const Grid = require('../../components/grid/grid.jsx');
-const TextArea = require('../../components/forms/textarea.jsx');
-const SubNavigation = require('../../components/subnavigation/subnavigation.jsx');
-const Select = require('../../components/forms/select.jsx');
 
 require('./components.scss');
 
 const Components = () => (
     <div className="components">
         <div className="inner">
-            <h1>Nav Bubbles</h1>
-            <div className="subnavigation">
-                <SubNavigation>
-                    <a href="">
-                        <li className="active">
-                            cats
-                        </li>
-                    </a>
-                    <a href="">
-                        <li>
-                            also cats
-                        </li>
-                    </a>
-                    <a href="">
-                        <li>
-                            not cats
-                        </li>
-                    </a>
-                </SubNavigation>
-            </div>
-            <h1>Grid</h1>
-            <Grid
-                showAvatar
-            />
             <h1>Button</h1>
-            <Button>I love buttons</Button>
+            <Button>I love button</Button>
             <h1>Form</h1>
-            <div className="form">
-                <Form>
-                    <Select
-                        label="Drop-down"
-                        required
-                        options={[
-                            {
-                                label: 'first option',
-                                value: 1
-                            },
-                            {
-                                label: 'second option',
-                                value: 2
-                            },
-                            {
-                                label: 'third option',
-                                value: 3
-                            }
-                        ]}
-                        name="name"
-                        value={1}
-                    />
-                    <Input
-                        label="Text input"
-                        required
-                        maxLength="30"
-                        name="test"
-                    />
-                    <TextArea
-                        label="Text area"
-                        name="textarea1"
-                        required
-                    />
-                </Form>
-            </div>
+            <Form>
+                <Input
+                    maxLength="30"
+                    name="test"
+                    type="text"
+                />
+            </Form>
             <h1>Box Component</h1>
             <Box
                 more="Cat Gifs"
