+++ src/views/components/components.scss
@@ -5,19 +5,6 @@
         margin: 0 0 10px 0;
     }
 
-    .subnavigation {
-        li {
-            background-color: $active-gray;
-            &.active {
-                background-color: $ui-blue;
-            }
-        }
-    }
-
-    .form {
-        width: 200px;
-    }
-
     .colors {
         span {
             display: inline-block;
