+++ src/views/preview/preview.jsx
@@ -5,7 +5,6 @@ const Page = require('../../components/page/www/page.jsx');
 const render = require('../../lib/render.jsx');
 
 const previewActions = require('../../redux/preview.js');
-const commentActions = require('../../redux/comments.js');
 
 const isSupportedBrowser = require('../../lib/supported-browser').default;
 const UnsupportedBrowser = require('./unsupported-browser.jsx');
@@ -17,7 +16,6 @@ if (isSupportedBrowser()) {
         document.getElementById('app'),
         {
             preview: previewActions.previewReducer,
-            comments: commentActions.commentsReducer,
             ...ProjectView.guiReducers
         },
         {
