+++ src/views/scratch_1.4/l10n.json
@@ -4,7 +4,7 @@
   "onePointFour.introNote": "{noteLabel} You can still share projects from 1.4 to the Scratch website. However, projects created in newer versions of Scratch cannot be opened in 1.4.",
   "onePointFour.downloads": "Downloads",
   "onePointFour.macTitle": "Mac OS X",
-  "onePointFour.macBody": "Compatible with Mac OSX 10.4 through 10.14",
+  "onePointFour.macBody": "Compatible with Mac OSX 10.4 or later",
   "onePointFour.windowsTitle": "Windows",
   "onePointFour.windowsBody": "Compatible with Windows 2000, XP, Vista, 7, and 8",
   "onePointFour.windowsNetworkInstaller": "installer",
