+++ src/views/studio/studio-info.jsx
@@ -3,11 +3,9 @@ import PropTypes from 'prop-types';
 import {useParams} from 'react-router-dom';
 import {connect} from 'react-redux';
 import Debug from './debug.jsx';
+import {getInfo, getRoles} from '../../redux/studio';
 
-import {selectUsername, selectToken} from '../../redux/session';
-import {getInfo, getRoles, selectCanEditInfo} from '../../redux/studio';
-
-const StudioInfo = ({username, studio, token, canEditInfo, onLoadInfo, onLoadRoles}) => {
+const StudioInfo = ({username, studio, token, onLoadInfo, onLoadRoles}) => {
     const {studioId} = useParams();
     
     useEffect(() => { // Load studio info after first render
@@ -25,16 +23,11 @@ const StudioInfo = ({username, studio, token, canEditInfo, onLoadInfo, onLoadRol
                 label="Studio Info"
                 data={studio}
             />
-            <Debug
-                label="Studio Info Permissions"
-                data={{canEditInfo}}
-            />
         </div>
     );
 };
 
 StudioInfo.propTypes = {
-    canEditInfo: PropTypes.bool,
     username: PropTypes.string,
     token: PropTypes.string,
     studio: PropTypes.shape({
@@ -45,12 +38,14 @@ StudioInfo.propTypes = {
 };
 
 export default connect(
-    state => ({
-        studio: state.studio,
-        username: selectUsername(state),
-        token: selectToken(state),
-        canEditInfo: selectCanEditInfo(state)
-    }),
+    state => {
+        const user = state.session.session.user;
+        return {
+            studio: state.studio,
+            username: user && user.username,
+            token: user && user.token
+        };
+    },
     dispatch => ({
         onLoadInfo: studioId => dispatch(getInfo(studioId)),
         onLoadRoles: (studioId, username, token) => dispatch(
