+++ src/views/studio/studio-projects.jsx
@@ -7,11 +7,10 @@ import {projectFetcher} from './lib/fetchers';
 import {projects} from './lib/redux-modules';
 import Debug from './debug.jsx';
 
-const {actions, selector: projectsSelector} = projects;
-import {selectCanAddProjects} from '../../redux/studio';
+const {actions, selector} = projects;
 
 const StudioProjects = ({
-    canAddProjects, items, error, loading, moreToLoad, onLoadMore
+    items, error, loading, moreToLoad, onLoadMore
 }) => {
     const {studioId} = useParams();
 
@@ -28,10 +27,6 @@ const StudioProjects = ({
                 label="Error"
                 data={error}
             />}
-            <Debug
-                label="Project Permissions"
-                data={{canAddProjects}}
-            />
             <div>
                 {items.map((item, index) =>
                     (<Debug
@@ -53,7 +48,6 @@ const StudioProjects = ({
 };
 
 StudioProjects.propTypes = {
-    canAddProjects: PropTypes.bool,
     items: PropTypes.array, // eslint-disable-line react/forbid-prop-types
     loading: PropTypes.bool,
     error: PropTypes.object, // eslint-disable-line react/forbid-prop-types
@@ -61,10 +55,7 @@ StudioProjects.propTypes = {
     onLoadMore: PropTypes.func
 };
 
-const mapStateToProps = state => ({
-    ...projectsSelector(state),
-    canAddProjects: selectCanAddProjects(state)
-});
+const mapStateToProps = state => selector(state);
 
 const mapDispatchToProps = dispatch => ({
     onLoadMore: (studioId, offset) => dispatch(
