+++ src/views/studio/studio.jsx
@@ -25,7 +25,6 @@ import {
 } from './lib/redux-modules';
 
 const {studioReducer} = require('../../redux/studio');
-const {commentsReducer} = require('../../redux/comments');
 
 const StudioShell = () => {
     const match = useRouteMatch();
@@ -76,7 +75,6 @@ render(
         [curators.key]: curators.reducer,
         [managers.key]: managers.reducer,
         [activity.key]: activity.reducer,
-        studio: studioReducer,
-        comments: commentsReducer
+        studio: studioReducer
     }
 );
