--- test/helpers/state-fixtures.json
@@ -1,51 +1,0 @@
-{
-    "studios": {
-        "isManager": {
-            "manager": true
-        },
-        "isCurator": {
-            "curator": true
-        },
-        "creator1": {
-            "owner": 1
-        },
-        "openToAll": {
-            "openToAll": true
-        }
-    },
-    "sessions": {
-        "user1Admin": {
-            "session": {
-                "user": {
-                    "id": 1
-                },
-                "permissions": {
-                    "admin": true
-                }
-            }
-        },
-        "user1Social": {
-            "session": {
-                "user": {
-                    "id": 1
-                },
-                "permissions": {
-                    "social": true
-                }
-            }
-        },
-        "user1": {
-            "session": {
-                "user": {
-                    "id": 1,
-                    "username": "user1-username",
-                    "token": "user1-token"
-                },
-                "permissions": {
-                    "admin": false,
-                    "social": false
-                }
-            }
-        }
-    }
-}
